Split up data.rs into submodule
